Refactoring: revert to previous check
mypy warned me about possibly comparing an int against an str, since
`constraint` might contain numbers or a '*'.  I thought I'd head it off
with this check, but that wasn't enough, so I found a different
workaround to placate mypy.

Anyway, the only position where '*' can appear is the last one, as this
is enforced by the regular expression in parse_python_requires().