Refactoring: introduce SourceFile
The name of the class is a bit misleading, but it makes sense in my
brain!  Source is an information source about supported Python versions,
it's more or less abstract.  SourceFile is a specific file used by a
source to declare Python versions (and other things).

The intent here is to allow Source to specify a glob so that it can
generate several SourceFiles.